React routing using v6.4+
A small project that implements the concept of react routing v6.4+. Many of the new features are explored here to gain knowledge about them. Of course, we can see the official document if we need more understanding, this is a simple play-and-learn project.

Topics that are explored here -
- V6.4+ way of declaring routes and nested routes, Also the Outlet to show the nested component
- Link/NavLink usecases
- New "loader" and "action" for data manipulation
- Important Hooks like - useLoaderData, useRouteLoaderData, useActionData, useRouterError, useNavigation, useFetcher etc
- exported functions of react-router-dom like - json, redirect, Await, defer etc

Looking forward to learn more.